Refresh the token before it expires
The recaptcha response (token) is only valid for 120 seconds after it's fetched. With the origiinal code, the token is only fetched once when the form is loaded; therefore, if a user submits the form more than 2 minutes after loading the page (which is a very common occurrence), the token would have expired and the recaptcha would systematically fail. 
This commit fixes the issue by refreshing the token every 100 seconds.